+++ src/BabDevMoneyBundle.php
@@ -5,7 +5,10 @@ namespace BabDev\MoneyBundle;
 use BabDev\MoneyBundle\DependencyInjection\BabDevMoneyExtension;
 use Doctrine\Bundle\DoctrineBundle\DependencyInjection\Compiler\DoctrineOrmMappingsPass;
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
-use Doctrine\ORM\UnitOfWork;
+use Doctrine\Bundle\MongoDBBundle\DependencyInjection\Compiler\DoctrineMongoDBMappingsPass;
+use Doctrine\Bundle\MongoDBBundle\DoctrineMongoDBBundle;
+use Doctrine\ODM\MongoDB\DocumentManager;
+use Doctrine\ORM\EntityManager;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
@@ -16,8 +19,13 @@ final class BabDevMoneyBundle extends Bundle
     {
         parent::build($container);
 
+        // Register ODM mappings if DoctrineMongoDBBundle and the ODM are installed
+        if (class_exists(DoctrineMongoDBBundle::class) && class_exists(DocumentManager::class)) {
+            $container->addCompilerPass(DoctrineMongoDBMappingsPass::createXmlMappingDriver([realpath(__DIR__.'../config/doctrine') => 'Money'], []));
+        }
+
         // Register ORM mappings if DoctrineBundle and the ORM are installed
-        if (class_exists(DoctrineBundle::class) && class_exists(UnitOfWork::class)) {
+        if (class_exists(DoctrineBundle::class) && class_exists(EntityManager::class)) {
             $container->addCompilerPass(DoctrineOrmMappingsPass::createXmlMappingDriver([realpath(__DIR__.'../config/doctrine') => 'Money']));
         }
     }